+++ ctp2_code/ui/aui_ctp2/c3_hypertextbox.cpp
@@ -230,7 +230,7 @@ AUI_ERRCODE c3_HyperTextBox::AddHyperStatics( const MBCHAR *hyperText )
 		hyperText = m_hyperText;
 	}
 
-	uint32 len = strlen( hyperText );
+	size_t len = strlen( hyperText );
 	if ( !len ) return AUI_ERRCODE_OK;
 
 	const MBCHAR *ptr = hyperText;
